Nav actionlib server, tests, and scout state machine
Actionlib server (and nodes in this repo interacting with it) now depend on drive_mode, which is a NAV_TYPE in common_names.h. New dispatch has not been tested.
Added tests to operations to test NavigationAlgo::changeInPosition. Tests have not been tested
Implemented scout state machine states used for the demo. Have not tested.